make download links cross-platform
